+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 -  Checkpoints now include model parameters, allowing for mismatches with the provided configuration file.
 - `accelerator` parameter now controls the accelerator (CPU, GPU, etc) that is used.
 - `devices` parameter controls the number of accelerators used.
-- `every_n_train_steps` parameter now controls the frequency of both validation epochs and model checkpointing during training.
+- `val_check_interval` parameter now controls the frequency of both validation epochs and model checkpointing during training.
 
 ### Changed
 
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Validation performance metrics are logged (and added to tensorboard) at the validation epoch, and training loss is logged at the end of training epoch, i.e. training and validation metrics are logged asynchronously.
 - Irrelevant warning messages on the console output and in the log file are no longer shown.
 - Nicely format logged warnings.
+- `every_n_train_steps` has been renamed to `val_check_interval` in accordance to the corresponding Pytorch Lightning parameter.
 
 ### Removed