+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- Don't try to assign non-existing output writer during eval mode.
+
 ## [3.4.0] - 2023-06-19
 
 ### Added